Separa metodo para salvar no db
