+++ index.html
@@ -84,6 +84,7 @@
     </div>
 
     <!-- You can also require other files to run in this process -->
+    <script src="./projectDAO.js"></script>
     <script src="./renderer.js"></script>
   </body>
 </html>
